refactor: Rename `lib/injectOutputOutputs.js`
