+++ lib/inject-output-outputs.js
