+++ lib/inject-output-outputs.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const injectOutputOutputs = require('./injectOutputOutputs');
+const injectOutputOutputs = require('./inject-output-outputs');
 
 describe('injectOutputOutputs', () => {
   it('adds cfn output for sfe output using Fn::Join', async () => {
