+++ lib/plugin.js
@@ -9,7 +9,7 @@ const login = require('./login');
 const logout = require('./logout');
 const wrap = require('./wrap');
 const injectLogsIamRole = require('./inject-logs-iam-role');
-const injectOutputOutputs = require('./injectOutputOutputs');
+const injectOutputOutputs = require('./inject-output-outputs');
 const wrapClean = require('./wrapClean');
 const getAppUids = require('./app-uids');
 const removeDestination = require('./removeDestination');
